## What does this change?

Housekeeping and turning on the cache functionality in eslint to speed up linting times.

Terribly un-random sample:

Before: W/ grunt-eslint 18, no cache

```
Execution Time (2016-09-07 13:59:37 UTC)
loading tasks         309ms  ▇ 2%
loading grunt-eslint  311ms  ▇ 2%
eslint:Gruntfile.js   650ms  ▇ 5%
eslint:grunt-configs  328ms  ▇ 2%
eslint:st...ascripts   4.7s  ▇▇▇▇▇▇▇ 34%
eslint:static/src      7.8s  ▇▇▇▇▇▇▇▇▇▇▇▇ 55%
Total 14.1s
```

After: W/ grunt-eslint 19, cache turned on

```
Execution Time (2016-09-07 16:09:54 UTC)
loading tasks         763ms  ▇▇▇▇ 16%
loading grunt-eslint  564ms  ▇▇▇ 12%
eslint:Gruntfile.js   923ms  ▇▇▇▇ 20%
eslint:grunt-configs  268ms  ▇▇ 6%
eslint:st...ascripts   1.3s  ▇▇▇▇▇▇ 27%
eslint:static/src     897ms  ▇▇▇▇ 19%
Total 4.7s
```

⚡ 

@guardian/dotcom-platform